Optimize Code, test=develop

+++ paddle/phi/kernels/fusion/xpu/fc_xpu_kernel.cc
@@ -119,7 +119,7 @@ void FcXPUKernel(const Context& ctx,
                  DenseTensor* out_max) {
   if (out_dtype == DataType::FLOAT32) {
     //FC_XPU_KERNEL_IMPL(float, float, float, float);
-    FC_XPU_KERNEL_IMPL(float, float, float, float);
+    FC_XPU_KERNEL_IMPL(float, float, float, int32_t);
   } else if (out_dtype == DataType::FLOAT16) {
     FC_XPU_KERNEL_IMPL(T, int16_t, dtype::float16, int16_t);
   } else {
